feat!: move old component to deprecated folder
