+++ src/components/deprecated/z-book-card-deprecated/index.stories.ts
@@ -2,8 +2,8 @@ import {Meta, StoryObj} from "@storybook/web-components";
 import {html} from "lit";
 import {styleMap} from "lit/directives/style-map.js";
 import {ZBookCardDeprecated} from ".";
-import {BookCardVariant} from "../../beans";
-import {CSSVarsArguments, getColorTokenArgConfig} from "../../utils/storybook-utils";
+import {BookCardVariant} from "../../../beans";
+import {CSSVarsArguments, getColorTokenArgConfig} from "../../../utils/storybook-utils";
 import "../z-button/index";
 import "../z-icon/index";
 import "../z-tag/index";
