+++ src/components/deprecated/z-book-card-deprecated/readme.md
@@ -45,7 +45,7 @@
 
 ### Depends on
 
-- [z-icon](../z-icon)
+- [z-icon](../../z-icon)
 
 ### Graph
 ```mermaid
