+++ src/components/z-book-card/readme.md
@@ -7,19 +7,21 @@
 
 ## Properties
 
-| Property        | Attribute        | Description                                            | Type                                                                                                                                     | Default                    |
-| --------------- | ---------------- | ------------------------------------------------------ | ---------------------------------------------------------------------------------------------------------------------------------------- | -------------------------- |
-| `adoption`      | `adoption`       | [optional] Show adoption badge                         | `boolean`                                                                                                                                | `false`                    |
-| `authors`       | `authors`        | [optional] Authors                                     | `string`                                                                                                                                 | `undefined`                |
-| `catalogUrl`    | `catalog-url`    | [optional] Show catalog link to correspondent resource | `string`                                                                                                                                 | `undefined`                |
-| `cover`         | `cover`          | Cover URL                                              | `string`                                                                                                                                 | `undefined`                |
-| `ebookUrl`      | `ebook-url`      | [optional] Show link to the ebook resource             | `string`                                                                                                                                 | `undefined`                |
-| `fallbackCover` | `fallback-cover` | [optional] Fallback cover URL                          | `string`                                                                                                                                 | `undefined`                |
-| `isbn`          | `isbn`           | [optional] Main ISBN                                   | `string`                                                                                                                                 | `undefined`                |
-| `operaTitle`    | `opera-title`    | [optional] Opera title                                 | `string`                                                                                                                                 | `undefined`                |
-| `tags`          | `tags`           | [optional] Tags                                        | `CardTag[] \| string`                                                                                                                    | `undefined`                |
-| `variant`       | `variant`        | Card variant: landscape, portrait                      | `BookCardVariant.COMPACT \| BookCardVariant.EXPANDED \| BookCardVariant.LANDSCAPE \| BookCardVariant.PORTRAIT \| BookCardVariant.SEARCH` | `BookCardVariant.PORTRAIT` |
-| `volumeTitle`   | `volume-title`   | [optional] Volume title                                | `string`                                                                                                                                 | `undefined`                |
+| Property         | Attribute         | Description                                            | Type                                                                                                                                     | Default                    |
+| ---------------- | ----------------- | ------------------------------------------------------ | ---------------------------------------------------------------------------------------------------------------------------------------- | -------------------------- |
+| `adoption`       | `adoption`        | [optional] Show adoption badge                         | `boolean`                                                                                                                                | `false`                    |
+| `annotated`      | `annotated`       | [optional] Annotated tag                               | `CardTag.ACTIVE \| CardTag.DISABLED`                                                                                                     | `undefined`                |
+| `authors`        | `authors`         | [optional] Authors                                     | `string`                                                                                                                                 | `undefined`                |
+| `catalogUrl`     | `catalog-url`     | [optional] Show catalog link to correspondent resource | `string`                                                                                                                                 | `undefined`                |
+| `cover`          | `cover`           | Cover URL                                              | `string`                                                                                                                                 | `undefined`                |
+| `ebookUrl`       | `ebook-url`       | [optional] Show link to the ebook resource             | `string`                                                                                                                                 | `undefined`                |
+| `edi`            | `edi`             | [optional] EDI tag                                     | `CardTag.ACTIVE \| CardTag.DISABLED`                                                                                                     | `undefined`                |
+| `fallbackCover`  | `fallback-cover`  | [optional] Fallback cover URL                          | `string`                                                                                                                                 | `undefined`                |
+| `isbn`           | `isbn`            | [optional] Main ISBN                                   | `string`                                                                                                                                 | `undefined`                |
+| `operaTitle`     | `opera-title`     | [optional] Opera title                                 | `string`                                                                                                                                 | `undefined`                |
+| `teacherVersion` | `teacher-version` | [optional] Annotated tag                               | `CardTag.ACTIVE \| CardTag.DISABLED`                                                                                                     | `undefined`                |
+| `variant`        | `variant`         | Card variant: landscape, portrait                      | `BookCardVariant.COMPACT \| BookCardVariant.EXPANDED \| BookCardVariant.LANDSCAPE \| BookCardVariant.PORTRAIT \| BookCardVariant.SEARCH` | `BookCardVariant.PORTRAIT` |
+| `volumeTitle`    | `volume-title`    | [optional] Volume title                                | `string`                                                                                                                                 | `undefined`                |
 
 
 ## Events
